chore(accounts receivable): Save float amounts to user QTY (#8631)

+++ app/Models/CustomerInvoiceItemDetails.php
@@ -379,7 +379,7 @@ class CustomerInvoiceItemDetails extends Model
         'VATAmountLocal' => 'float',
         'VATAmountRpt' => 'float',
         'VATApplicableOn' => 'integer',
-        'userQty' => 'integer'
+        'userQty' => 'float'
     ];
 
     /**
